get rid of deprection warnings in EventFilter/SiStripRawToDigi
